Phase 1A PR2: extract MultiSelect/DraggableSlider into common; update imports; add unit tests; docs
