+++ components/list-builder/common/multi-select.tsx
@@ -1,0 +1,107 @@
+"use client"
+
+import React, { useState } from "react"
+import { Label } from "@/components/ui/label"
+import { Tooltip, TooltipContent, TooltipTrigger } from "@/components/ui/tooltip"
+import { Checkbox } from "@/components/ui/checkbox"
+import { Badge } from "@/components/ui/badge"
+import { Button } from "@/components/ui/button"
+import { ChevronDown, ChevronUp, Info, X } from "lucide-react"
+
+export interface MultiSelectOption {
+  value: string
+  label: string
+}
+
+export interface MultiSelectProps {
+  label: string
+  options: MultiSelectOption[]
+  selected: string[]
+  onChange: (selected: string[]) => void
+  icon?: React.ReactNode
+  tooltip?: string
+  placeholder?: string
+}
+
+export function MultiSelect({ label, options, selected, onChange, icon, tooltip, placeholder }: MultiSelectProps) {
+  const [isOpen, setIsOpen] = useState(false)
+
+  const handleToggle = (value: string) => {
+    const newSelected = selected.includes(value) ? selected.filter((item) => item !== value) : [...selected, value]
+    onChange(newSelected)
+  }
+
+  const handleRemove = (value: string) => {
+    onChange(selected.filter((item) => item !== value))
+  }
+
+  return (
+    <div className="space-y-3">
+      <div className="flex items-center gap-2">
+        {icon}
+        <Label className="font-medium">{label}</Label>
+        {tooltip && (
+          <Tooltip>
+            <TooltipTrigger>
+              <Info className="h-4 w-4 text-gray-400" />
+            </TooltipTrigger>
+            <TooltipContent>
+              <p className="max-w-xs">{tooltip}</p>
+            </TooltipContent>
+          </Tooltip>
+        )}
+      </div>
+
+      {/* Selected Items */}
+      {selected.length > 0 && (
+        <div className="flex flex-wrap gap-2">
+          {selected.map((value) => {
+            const option = options.find((opt) => opt.value === value)
+            return (
+              <Badge key={value} variant="secondary" className="bg-yellow-100 text-yellow-800 hover:bg-yellow-200">
+                {option?.label || value}
+                <Button
+                  variant="ghost"
+                  size="sm"
+                  className="h-4 w-4 p-0 ml-1 hover:bg-yellow-300"
+                  onClick={() => handleRemove(value)}
+                >
+                  <X className="h-3 w-3" />
+                </Button>
+              </Badge>
+            )
+          })}
+        </div>
+      )}
+
+      {/* Dropdown */}
+      <div className="relative">
+        <Button variant="outline" onClick={() => setIsOpen(!isOpen)} className="w-full justify-between">
+          <span className="text-gray-500">
+            {selected.length > 0 ? `${selected.length} selected` : placeholder || `Select ${label}`}
+          </span>
+          {isOpen ? <ChevronUp className="h-4 w-4" /> : <ChevronDown className="h-4 w-4" />}
+        </Button>
+
+        {isOpen && (
+          <div className="absolute z-10 w-full mt-1 bg-white dark:bg-gray-800 border border-gray-200 dark:border-gray-700 rounded-md shadow-lg max-h-60 overflow-auto">
+            {options.map((option) => (
+              <div
+                key={option.value}
+                className="flex items-center space-x-2 px-3 py-2 hover:bg-gray-100 dark:hover:bg-gray-700 cursor-pointer"
+                onClick={() => handleToggle(option.value)}
+              >
+                <Checkbox
+                  checked={selected.includes(option.value)}
+                  onCheckedChange={() => handleToggle(option.value)}
+                  onClick={(e) => e.stopPropagation()}
+                />
+                <span className="text-sm">{option.label}</span>
+              </div>
+            ))}
+          </div>
+        )}
+      </div>
+    </div>
+  )
+}
