+++ components/list-builder/demographics-filters.tsx
@@ -67,278 +67,17 @@ import {
   CHILDREN_AGE_RANGES,
 } from "@/data/demographics"
 
+import { DraggableSlider } from "@/components/list-builder/common/draggable-slider"
+import { MultiSelect } from "@/components/list-builder/common/multi-select"
+
 interface DemographicsFiltersProps {
   criteria: DemographicsCriteria
   onUpdate: (values: Partial<DemographicsCriteria>) => void
 }
 
-interface DraggableSliderProps {
-  label: string
-  value: number[]
-  min: number
-  max: number
-  step: number
-  formatValue: (value: number) => string
-  onChange: (value: number[]) => void
-  ariaLabel: string
-  icon?: React.ReactNode
-  tooltip?: string
-}
+// Components moved to '@/components/list-builder/common/*'
 
-// Option constants moved to '@/data/demographics'
-
-function DraggableSlider({
-  label,
-  value,
-  min,
-  max,
-  step,
-  formatValue,
-  onChange,
-  ariaLabel,
-  icon,
-  tooltip,
-}: DraggableSliderProps) {
-  const [isDragging, setIsDragging] = useState<"min" | "max" | null>(null)
-  const sliderRef = useRef<HTMLDivElement>(null)
-
-  const getSliderBounds = useCallback(() => {
-    if (!sliderRef.current) return { left: 0, width: 0 }
-    const rect = sliderRef.current.getBoundingClientRect()
-    return { left: rect.left, width: rect.width }
-  }, [])
-
-  const valueToPosition = useCallback(
-    (val: number) => {
-      return ((val - min) / (max - min)) * 100
-    },
-    [min, max],
-  )
-
-  const positionToValue = useCallback(
-    (position: number, bounds: { left: number; width: number }) => {
-      const percentage = Math.max(0, Math.min(100, (position / bounds.width) * 100))
-      const rawValue = min + (percentage / 100) * (max - min)
-      return Math.round(rawValue / step) * step
-    },
-    [min, max, step],
-  )
-
-  const handleMouseDown = useCallback(
-    (handle: "min" | "max") => (e: React.MouseEvent) => {
-      e.preventDefault()
-      setIsDragging(handle)
-      document.body.style.userSelect = "none"
-    },
-    [],
-  )
-
-  const handleMouseMove = useCallback(
-    (e: MouseEvent) => {
-      if (!isDragging || !sliderRef.current) return
-
-      const bounds = getSliderBounds()
-      const clientX = e.clientX - bounds.left
-      const newValue = positionToValue(clientX, bounds)
-
-      if (isDragging === "min") {
-        const newMin = Math.max(min, Math.min(newValue, value[1] - step))
-        onChange([newMin, value[1]])
-      } else {
-        const newMax = Math.min(max, Math.max(newValue, value[0] + step))
-        onChange([value[0], newMax])
-      }
-    },
-    [isDragging, value, onChange, getSliderBounds, positionToValue, min, max, step],
-  )
-
-  const handleMouseUp = useCallback(() => {
-    setIsDragging(null)
-    document.body.style.userSelect = ""
-  }, [])
-
-  useEffect(() => {
-    if (isDragging) {
-      document.addEventListener("mousemove", handleMouseMove)
-      document.addEventListener("mouseup", handleMouseUp)
-      return () => {
-        document.removeEventListener("mousemove", handleMouseMove)
-        document.removeEventListener("mouseup", handleMouseUp)
-      }
-    }
-  }, [isDragging, handleMouseMove, handleMouseUp])
-
-  return (
-    <div className="space-y-4">
-      <div className="flex items-center gap-2">
-        {icon}
-        <Label className="font-medium">{label}</Label>
-        {tooltip && (
-          <Tooltip>
-            <TooltipTrigger>
-              <Info className="h-4 w-4 text-gray-400" />
-            </TooltipTrigger>
-            <TooltipContent>
-              <p className="max-w-xs">{tooltip}</p>
-            </TooltipContent>
-          </Tooltip>
-        )}
-      </div>
-
-      <div
-        ref={sliderRef}
-        className="relative h-6 bg-gray-200 dark:bg-gray-700 rounded-full cursor-pointer"
-        role="slider"
-        aria-label={ariaLabel}
-        aria-valuemin={min}
-        aria-valuemax={max}
-        aria-valuenow={value[1]}
-        aria-valuetext={`${formatValue(value[0])} to ${formatValue(value[1])}`}
-        tabIndex={-1}
-      >
-        <div
-          className="absolute h-full bg-yellow-500 rounded-full transition-all duration-150"
-          style={{
-            left: `${valueToPosition(value[0])}%`,
-            width: `${valueToPosition(value[1]) - valueToPosition(value[0])}%`,
-          }}
-        />
-
-        <div
-          className={`absolute top-1/2 w-6 h-6 bg-white border-2 border-yellow-500 rounded-full shadow-md cursor-grab transform -translate-y-1/2 -translate-x-1/2 transition-all duration-150 hover:scale-110 focus:outline-none focus:ring-2 focus:ring-yellow-500 focus:ring-offset-2 ${
-            isDragging === "min" ? "cursor-grabbing scale-110 ring-2 ring-yellow-500 ring-offset-2" : ""
-          }`}
-          style={{ left: `${valueToPosition(value[0])}%` }}
-          onMouseDown={handleMouseDown("min")}
-          role="slider"
-          aria-label={`${label} minimum value`}
-          aria-valuemin={min}
-          aria-valuemax={value[1] - step}
-          aria-valuenow={value[0]}
-          aria-valuetext={formatValue(value[0])}
-          tabIndex={0}
-        >
-          <div className="absolute inset-0 flex items-center justify-center">
-            <div className="w-2 h-2 bg-yellow-500 rounded-full" />
-          </div>
-        </div>
-
-        <div
-          className={`absolute top-1/2 w-6 h-6 bg-white border-2 border-yellow-500 rounded-full shadow-md cursor-grab transform -translate-y-1/2 -translate-x-1/2 transition-all duration-150 hover:scale-110 focus:outline-none focus:ring-2 focus:ring-yellow-500 focus:ring-offset-2 ${
-            isDragging === "max" ? "cursor-grabbing scale-110 ring-2 ring-yellow-500 ring-offset-2" : ""
-          }`}
-          style={{ left: `${valueToPosition(value[1])}%` }}
-          onMouseDown={handleMouseDown("max")}
-          role="slider"
-          aria-label={`${label} maximum value`}
-          aria-valuemin={value[0] + step}
-          aria-valuemax={max}
-          aria-valuenow={value[1]}
-          aria-valuetext={formatValue(value[1])}
-          tabIndex={0}
-        >
-          <div className="absolute inset-0 flex items-center justify-center">
-            <div className="w-2 h-2 bg-yellow-500 rounded-full" />
-          </div>
-        </div>
-      </div>
-
-      <div className="flex justify-between text-sm text-gray-600 dark:text-gray-400">
-        <span className="font-medium">{formatValue(value[0])}</span>
-        <span className="font-medium">{formatValue(value[1])}</span>
-      </div>
-    </div>
-  )
-}
-
-interface MultiSelectProps {
-  label: string
-  options: { value: string; label: string }[]
-  selected: string[]
-  onChange: (selected: string[]) => void
-  icon?: React.ReactNode
-  tooltip?: string
-  placeholder?: string
-}
-
-function MultiSelect({ label, options, selected, onChange, icon, tooltip, placeholder }: MultiSelectProps) {
-  const [isOpen, setIsOpen] = useState(false)
-
-  const handleToggle = (value: string) => {
-    const newSelected = selected.includes(value) ? selected.filter((item) => item !== value) : [...selected, value]
-    onChange(newSelected)
-  }
-
-  const handleRemove = (value: string) => {
-    onChange(selected.filter((item) => item !== value))
-  }
-
-  return (
-    <div className="space-y-3">
-      <div className="flex items-center gap-2">
-        {icon}
-        <Label className="font-medium">{label}</Label>
-        {tooltip && (
-          <Tooltip>
-            <TooltipTrigger>
-              <Info className="h-4 w-4 text-gray-400" />
-            </TooltipTrigger>
-            <TooltipContent>
-              <p className="max-w-xs">{tooltip}</p>
-            </TooltipContent>
-          </Tooltip>
-        )}
-      </div>
-
-      {/* Selected Items */}
-      {selected.length > 0 && (
-        <div className="flex flex-wrap gap-2">
-          {selected.map((value) => {
-            const option = options.find((opt) => opt.value === value)
-            return (
-              <Badge key={value} variant="secondary" className="bg-yellow-100 text-yellow-800 hover:bg-yellow-200">
-                {option?.label || value}
-                <Button
-                  variant="ghost"
-                  size="sm"
-                  className="h-4 w-4 p-0 ml-1 hover:bg-yellow-300"
-                  onClick={() => handleRemove(value)}
-                >
-                  <X className="h-3 w-3" />
-                </Button>
-              </Badge>
-            )
-          })}
-        </div>
-      )}
-
-      {/* Dropdown */}
-      <div className="relative">
-        <Button variant="outline" onClick={() => setIsOpen(!isOpen)} className="w-full justify-between">
-          <span className="text-gray-500">
-            {selected.length > 0 ? `${selected.length} selected` : placeholder || `Select ${label}`}
-          </span>
-          {isOpen ? <ChevronUp className="h-4 w-4" /> : <ChevronDown className="h-4 w-4" />}
-        </Button>
-
-        {isOpen && (
-          <div className="absolute z-10 w-full mt-1 bg-white dark:bg-gray-800 border border-gray-200 dark:border-gray-700 rounded-md shadow-lg max-h-60 overflow-auto">
-            {options.map((option) => (
-              <div
-                key={option.value}
-                className="flex items-center space-x-2 px-3 py-2 hover:bg-gray-100 dark:hover:bg-gray-700 cursor-pointer"
-                onClick={() => handleToggle(option.value)}
-              >
-                <Checkbox checked={selected.includes(option.value)} onChange={() => handleToggle(option.value)} />
-                <span className="text-sm">{option.label}</span>
-              </div>
-            ))}
-          </div>
-        )}
-      </div>
-    </div>
-  )
-}
+// Components moved to '@/components/list-builder/common/*'
 
 export function DemographicsFilters({ criteria, onUpdate }: DemographicsFiltersProps) {
   const [showPresetDialog, setShowPresetDialog] = useState(false)
